+++ src/api/admin.rs
@@ -1,0 +1,4 @@
+use openapi::apis::adminschnittstellen_autoren::*;
+use openapi::apis::adminschnittstellen_dokumente::*;
+use openapi::apis::adminschnittstellen_enumerations::*;
+use openapi::apis::adminschnittstellen_gremien::*;
